Soft-Refactor Geiger Counters (#615)
# Description

This refactors Geiger Counters so that their behavior of "Only making
sound to a person holding them" is no longer hardcoded. The
GeigerCounterComponent now can define how loud it ticks, how far away
people can hear the ticks, and whether it plays only for the person
holding it or for anyone nearby. This PR partially fulfills one of the
"Nice To Have" features requested for
https://github.com/Simple-Station/Einstein-Engines/pull/341 by making it
possible to create stationary radiation alarm objects. It also serves as
a substantial quality of life improvement for Engineering and Science
crew, since it's now possible to place an active Geiger counter in the
artifact lab, and then be able to audibly hear if the lab becomes
radioactive due to an artifact.

<details><summary><h1>Media</h1></summary>
<p>



https://github.com/user-attachments/assets/74122135-7345-4995-bb0e-d1216e1d53b6



https://github.com/user-attachments/assets/de79db6f-e1c1-471f-88b5-0a47ff4bfa16


</p>
</details>

# Changelog

:cl:
- add: Geiger Counters other than ones installed in Hardsuits now
generate an audible sound when active and exposed to radiation.
- add: Wall mounted geiger counters have been added to the game.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>