Rename Mantis (#709)
"Mantis" was going to be the original name, but the doubling as
detective lead me to needing to add the qualifier "forensic" to make
that part clearer. The "Mantis" was already there to imply the psionic
part.

For "psionic mantis" defenders, I propose an alternative renaming
scheme:
Mystagogue -> Psionic Mystagogue
Mantis -> Psionic Mantis
Cyborg -> Robotic Cyborg
Chaplain -> Religious Chaplain

Players either already know what a mantis is, or they've got little
enough playtime that there's still some intrigue left in the setting and
it's something they can learn by observation. Weird names are mostly
restricted to one dept here and a tinge harder time understanding it is
what I'd call "Mystery" rather than something that's undesirable here.
The mechanic the mantis interacts with - psionics - is also somewhat
hidden and esoteric.

# Media

![image](https://github.com/user-attachments/assets/8706c0cd-97a3-4ed6-a0e4-e23012a461a7)


![image](https://github.com/user-attachments/assets/7e6296fb-52ca-4d4e-b40d-2e166520d4b5)


# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Rane
- tweak: Renamed "Psionic Mantis" to "Mantis", as it was originally
going to be called.